Lab Card Demo Bot/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
Flask==2.0.1
future==0.18.2
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyJWT==2.1.0
python-dateutil==2.8.2
requests==2.26.0
requests-toolbelt==0.9.1
six==1.16.0
smartsheet-python-sdk==2.105.1
urllib3==1.26.7
webexteamssdk==1.6
Werkzeug==2.0.1
